### 2020-03-04 #2138 @whedon
[PRE REVIEW]: py-pde: A Python package for solving partial differential equations
**Submitting author:** @david-zwicker (<a href="http://orcid.org/0000-0002-3909-3334">David Zwicker</a>)
**Repository:** <a href="https://github.com/zwicker-group/py-pde" target ="_blank">https://github.com/zwicker-group/py-pde</a>
**Version:** v0.2
**Reviewers:** @celliern, @@mstimberg
