+++ tests/tests_run.sh
@@ -7,9 +7,9 @@ if [ ! -z $1 ]
 then 
 	# test pattern was specified 
 	echo 'Run unittests with pattern '$1
-	python3 -m pytest -k "$1" . ..
+	python3 -m pytest -rs -k "$1" . ..
 else
 	# test pattern was not specified
 	echo 'Run all unittests'
-    python3 -m pytest . ..
+    python3 -m pytest -rs . ..
 fi
