### 2020-03-05 dc242fe
Removed analysis package since it was not very useful
Changed logger names to be more useful
