+++ pde/common.py
@@ -12,4 +12,3 @@ from .storage import *  # @UnusedWildImport
 from .trackers import *  # @UnusedWildImport
 from .visualization import *  # @UnusedWildImport
 from .controller import Controller  # @UnusedImport
-from .analysis import *  # @UnusedWildImport
