### 2020-03-05 c70199c
Larger refactoring
Controller has been moved to solvers
Common module has been removed and is now available in the root package
Improved documentation structure
