+++ docs/source/run_autodoc.py
@@ -59,6 +59,7 @@ def main():
     for path in glob.glob(f'{OUTPUT_PATH}/*.rst'):
         logging.info('Patch file `%s`', path)
         replace_in_file(path, {'Submodules\n----------\n\n': ''})
+        replace_in_file(path, {'Subpackages\n-----------\n\n': ''})
     
     
 if __name__ == '__main__':
