+++ pde/__init__.py
@@ -3,7 +3,14 @@ The py-pde package provides classes and methods for solving partial differential
 equations.
 """
 
-from .tools.misc import environment
-
-
 __version__ = "0.2.2"
+
+from .fields import *  # @UnusedWildImport
+from .grids import *  # @UnusedWildImport
+from .pdes import *  # @UnusedWildImport
+from .solvers import *  # @UnusedWildImport
+from .storage import *  # @UnusedWildImport
+from .trackers import *  # @UnusedWildImport
+from .visualization import *  # @UnusedWildImport
+
+from .tools.misc import environment
