+++ pde/solvers/tests/test_explicit.py
@@ -8,8 +8,7 @@ import numpy as np
 from ...fields import ScalarField
 from ...grids import UnitGrid
 from ...pdes import DiffusionPDE
-from ...controller import Controller
-from .. import ExplicitSolver
+from .. import Controller, ExplicitSolver
 
 
 
