### 2020-03-05 53ba08d
Bumped version number to 0.3
