+++ pde/__init__.py
@@ -3,7 +3,7 @@ The py-pde package provides classes and methods for solving partial differential
 equations.
 """
 
-__version__ = "0.2.2"
+__version__ = "0.3"
 
 from .fields import *  # @UnusedWildImport
 from .grids import *  # @UnusedWildImport
