+++ setup.py
@@ -8,7 +8,7 @@ setup(
   package_data={"pde": ["py.typed"]},
   packages = find_packages(),
   zip_safe=False,  # this is required for mypy to find the py.typed file
-  version = '0.2.2',
+  version = '0.3',
   license='MIT',
   description = 'Python package for solving partial differential equations',
   long_description=long_description,
@@ -16,7 +16,7 @@ setup(
   author = 'David Zwicker',
   author_email = 'david.zwicker@ds.mpg.de',
   url = 'https://github.com/zwicker-group/py-pde',
-  download_url = 'https://github.com/zwicker-group/py-pde/archive/v0.2.2.tar.gz',
+  download_url = 'https://github.com/zwicker-group/py-pde/archive/v0.3.tar.gz',
   keywords = ['pdes', 'partial-differential-equations', 'dynamical-systems'],
   python_requires='>=3.6',
   install_requires=['matplotlib',
