+++ docs/source/conf.py
@@ -65,6 +65,7 @@ extensions = [
     'sphinx.ext.napoleon',
 #     'sphinx_autodoc_annotation',
     'sphinx_simplify_typehints',
+    'sphinx.ext.inheritance_diagram',
 ]
 
 # Add any paths that contain templates here, relative to this directory.
