### 2020-03-09 6ca47cb
Improved documentation style
