+++ docs/source/index.rst
@@ -21,7 +21,7 @@ Contents
 --------
 
 .. toctree::
-    :maxdepth: 4
+    :maxdepth: 2
 
     installation
     quickstart/quickstart
