### 2020-03-10 8188a0f
Enabled compression and intermediate storage in FileStorage
This is an effort to store results from large simulations more
efficiently. Moreover, this should enable the reading of partial
simulation data.
