### 2020-03-10 403dfc3
Fixed writing to FileStorage with keep_opened == False
