### 2020-03-10 da615c8
Moved improved parameter class from other package
