+++ pde/tools/parameters.py
@@ -1,37 +1,44 @@
 '''
-Helper classes for dealing with instance parameters
+This module provides infrastructure for managing classes with parameters. One
+aim is to allow easy management of inheritance.
+
 
 .. codeauthor:: David Zwicker <david.zwicker@ds.mpg.de>
 '''
 
 
 import logging
-from typing import Sequence, Dict, Any  # @UnusedImport
+from collections import OrderedDict
+from typing import Sequence, Dict, Any, Set  # @UnusedImport
 
-from ..tools.misc import import_class, hybridmethod
+from pde.tools.misc import import_class, hybridmethod, classproperty
 
 
 
 class Parameter():
-    """ class representing information about parameters """
+    """ class representing a single parameter """
     
-    def __init__(self, name: str, default_value=None,
-                 cls=object, description: str = '',
-                 deprecated: bool = False):
-        """ initialize a single parameter
+    def __init__(self, name: str,
+                 default_value=None,
+                 cls=object,
+                 description: str = ''):
+        """ initialize a parameter 
         
         Args:
-            name (str): The name of the parameter
-            default_value: The default value
-            cls: The type of the parameter also used to convert it
-            description (str): A description of the parameter
-            bool (flag): Flag stating whether this parameter is deprecated
+            name (str):
+                The name of the parameter
+            default_value:
+                The default value
+            cls:
+                The type of the parameter, which is used for conversion
+            description (str):
+                A string describing the impact of this parameter. This
+                description appears in the parameter help
         """
         self.name = name
         self.default_value = default_value
         self.cls = cls
         self.description = description
-        self.deprecated = deprecated
         if cls is not object and cls(default_value) != default_value:
             logging.warning('Default value `%s` does not seem to be of type '
                             '`%s`', name, cls.__name__)
@@ -73,14 +80,17 @@ class Parameter():
         
         
     def convert(self, value=None):
-        """ converts a value into the correct type for this parameter.
+        """ converts a `value` into the correct type for this parameter. If
+        `value` is not given, the default value is converted.
         
-        If value is not given, the default value is converted. Note that this
-        does not necessarily copy the values, which could lead to unexpected
-        effects where the default value is changed by an instance.
+        Note that this does not make a copy of the values, which could lead to
+        unexpected effects where the default value is changed by an instance.
         
         Args:
-            value: The value which will be converted
+            value: The value to convert
+        
+        Returns:
+            The converted value, which is of type `self.cls`
         """ 
         if value is None:
             value = self.default_value
@@ -92,33 +102,60 @@ class Parameter():
 
 
 
+class DeprecatedParameter(Parameter):
+    """ a parameter that can still be used normally but is deprecated """
+    pass
+
+
+
+class ObsoleteParameter():
+    """ parameter that was defined on a parent class, but does not apply to
+    child class anymore. This class can be used on child classes to signal that
+    a parameter from the parent class must not be inherited. """
+    
+    
+    def __init__(self, name: str):
+        """ 
+        Args:
+            name (str):
+                The name of the parameter
+        """
+        self.name = name
+
+
+
 class Parameterized():
-    """ manages a dictionary of parameters assigned to classes """
+    """ a mixin that manages the parameters of a class """
 
     parameters_default: Sequence[Parameter] = []
     _subclasses: Dict[str, Any] = {}
 
 
-    def __init__(self, parameters=None, check_validity=True):
+    def __init__(self, parameters: Dict[str, Any] = None,
+                 check_validity: bool = True):
         """ initialize the object with optional parameters that overwrite the
         default behavior
         
         Args:
             parameters (dict):
-                A dictionary of parameters overwriting the defaults
+                A dictionary of parameters to change the defaults
             check_validity (bool):
-                Determines whether an error is raised if there are keys in
[…]
