+++ pde/tools/tests/test_parameters.py
@@ -1,48 +1,106 @@
 '''
-Check the parameters module
-
 .. codeauthor:: David Zwicker <david.zwicker@ds.mpg.de>
 '''
 
-from ..parameters import Parameter, Parameterized
+import itertools
+
+import pytest
+
+from ..parameters import (Parameter, DeprecatedParameter, ObsoleteParameter,
+                          Parameterized, get_all_parameters)
 
 
 
-def test_parameter_class():
-    """ test Parameter class """
-    p = Parameter(name='int', default_value=1, cls=int, description='test',
-                  deprecated=False)
+def test_parameters():
+    """ test mixing Parameterized """
     
-    assert p.name == 'int'
-    assert p.convert(1.1) == 1
-    
-    q = Parameter('new')
-    q.__setstate__(p.__getstate__())
-    assert p.__dict__ == q.__dict__
-    
-    
+    class Test1(Parameterized):
+        parameters_default = [Parameter('a', 1, int, "help")]
+        
+        
+    t = Test1()
+    assert t.parameters['a'] == 1
+    assert t.get_parameter_default('a') == 1
 
-def test_parameterized_class(capsys):
-    """ test Parameter class """
+    t = Test1(parameters={'a': 2})
+    assert t.parameters['a'] == 2
+    assert t.get_parameter_default('a') == 1
     
-    class Test(Parameterized):
-        parameters_default = [Parameter(name='int', default_value=1, cls=int,
-                                        description='test', deprecated=False)]
+    with pytest.raises(ValueError):
+        t = Test1(parameters={'b': 1})
+    t = Test1(parameters={'b': 1}, check_validity=False)
+    assert t.parameters['a'] == 1
+    assert t.parameters['b'] == 1
     
-    Test.show_parameters(description=True)
-    captured = capsys.readouterr()
-    assert 'int' in captured.out
-    assert '1' in captured.out
     
-    t = Test()
-    assert t.parameters['int'] == 1
+    class Test2(Test1):
+        # also test conversion of default parameters
+        parameters_default = [Parameter('b', '2', int, "help")]
+        
+    t = Test2()
+    assert t.parameters['a'] == 1
+    assert t.parameters['b'] == 2
     
-    t = Test(parameters={'int': 2})
-    assert t.parameters['int'] == 2
-    assert t.get_parameter_default('int') == 1
-    t.show_parameters(description=True)
-    captured = capsys.readouterr()
-    assert 'int' in captured.out
-    assert '2' in captured.out
+    t = Test2(parameters={'a': 10, 'b': 20})
+    assert t.parameters['a'] == 10
+    assert t.parameters['b'] == 20
+    assert t.get_parameter_default('a') == 1
+    assert t.get_parameter_default('b') == '2'
+    with pytest.raises(KeyError):
+        t.get_parameter_default('c')
+        
+    class Test3(Test2):
+        # test overwriting defaults
+        parameters_default = [Parameter('a', 3), Parameter('c', 4)]
+        
+    t = Test3()
+    assert t.parameters['a'] == 3
+    assert t.get_parameter_default('a') == 3
+    assert set(t.parameters.keys()) == {'a', 'b', 'c'}
+        
+        
+        
+def test_parameter_help(capsys):
+    """ test how parameters are shown """
+    class Test1(Parameterized):
+        parameters_default = [DeprecatedParameter('a', 1, int, "random string")]
+        
+    class Test2(Test1):
+        parameters_default = [Parameter('b', 2, int, "another word")]
+        
+    t = Test2()
+    for flags in itertools.combinations_with_replacement([True, False], 3):
+        Test2.show_parameters(*flags)
[…]
