+++ pde/tools/cuboid.py
@@ -1,4 +1,7 @@
 '''
+This module defines the :class:`Cuboid` class, which represents an n-dimensional
+cuboid that is aligned with the axes of a Cartesian coordinate system.
+
 .. codeauthor:: David Zwicker <david.zwicker@ds.mpg.de>
 '''
 
