### 2020-03-10 #2138 @celliern
@xuanxu Hello! If you need it, I would be happy to review this work. I have been recently published in JOSS for a similar software ([here](https://joss.theoj.org/papers/10.21105/joss.01356)) and have some skill in Python and PDE modeling. Do not feel obliged to, if you find a more relevant reviewer or if you find both works being to close to each other.
