### 2020-03-10 #2158 @whedon
[REVIEW]: py-pde: A Python package for solving partial differential equations
**Submitting author:** @david-zwicker (<a href="http://orcid.org/0000-0002-3909-3334">David Zwicker</a>)
**Repository:** <a href="https://github.com/zwicker-group/py-pde" target ="_blank">https://github.com/zwicker-group/py-pde</a>
**Version:** v0.5

@celliern & @@mstimberg, please carry out your review in this issue by updating the checklist below. If you cannot edit the checklist please:
