### 2020-03-10 #2158 @danielskatz
This is a problem with a reference in the .md and .bib files - the .md file is trying to reference something that is not in the bib file (or the bib file has a problem that leads to it not being parsed correctly)
