### 2020-03-11 #2158 @mstimberg
@xuanxu I cannot edit the checklist and I did not receive an invite – is this because this is my first review for JOSS and I have to do something else first? Or is it because of the earlier `@@mstimberg` issue?
