### 2020-03-11 b2d60c0
Improved field type detection in storage classes
