+++ pde/storage/tests/test_memory.py
@@ -2,11 +2,12 @@
 .. codeauthor:: David Zwicker <david.zwicker@ds.mpg.de>
 '''
 
+import pytest
 import numpy as np
 
 from .. import MemoryStorage
 from ...grids import UnitGrid
-from ...fields import ScalarField
+from ...fields import ScalarField, VectorField, Tensor2Field, FieldCollection
 
 
             
@@ -24,3 +25,36 @@ def test_memory_storage():
     
     s3 = MemoryStorage.from_collection([s1, s2])
     np.testing.assert_allclose(np.ravel(s3.data), np.arange(4))
+
+
+
+def test_field_type_guessing():
+    """ test the ability to guess the field type """
+    for cls in [ScalarField, VectorField, Tensor2Field]:
+        grid = UnitGrid([3])
+        field = cls.random_normal(grid)
+        s = MemoryStorage()
+        s.start_writing(field)
+        s.append(field.data, 0)
+        s.append(field.data, 1)
+        
+        # delete information
+        s._field = None
+        s.info = {}
+        
+        assert len(s) == 2
+        assert s[0] == field
+        
+    field = FieldCollection([ScalarField(grid), VectorField(grid)])
+    s = MemoryStorage()
+    s.start_writing(field)
+    s.append(field.data, 0)
+    
+    # delete information
+    s._field = None
+    s.info = {}
+    
+    with pytest.raises(RuntimeError):
+        s[0]
+
+    
