+++ tests/mypy.ini
@@ -5,9 +5,6 @@ warn_return_any = True
 warn_unused_configs = True
 warn_unused_ignores = True
 
-[mypy-droplets.contrib.*]
-ignore_errors = True
-
 [mypy-h5py.*]
 ignore_missing_imports = True
 
