### 2020-03-12 6697919
Moved spectral density calculation to test module
Removed unnecessary __del__ from FileStorage
