### 2020-03-13 #2158 @celliern
First, I wanted to say that I'm impressed ! The library is clean and well
organized and I will definitively add it in my research toolbox. It also fit
well the Python scientific ecosystem, especially compared with my work () that
has not be designed to easily write tensor-based PDE equation. This is a
welcome approach as this is a very common task among a wide range of
scientific topics.

The features offered by py-pde are multiple and the dependencies despite
very narrowed mandatory dependencies (all of them being in the standard
scientific stack and shipped with distributions like Anaconda). This should
guarantee an easy deployment on clusters.

All tests have run with a fresh conda environment with py=3.6 on my computer.

I still have to play a bit with it to ensure the functional claims of the
software. During this time, I have some optional (but welcome) improvement that
I could suggest :

## Packaging improvement

The optional dependencies could be listed in the `setup.py` ([see in
the setuptools doc](https://setuptools.readthedocs.io/en/latest/
setuptools.html#declaring-extras-optional-features-with-their-own-
dependencies)) to make the process more automatic for the user.

## Documentation improvement

Some information are available in the paper may be added also in the
documentation: the method used for the operator discretization (finite
difference), the temporal solvers available for the explicit as well as
for the implicit ones. In general, a "advanced user documentation" could
highly improve the adoption of the software : more detail on how to implement
custom PDE (with or without numba), some info on the internals (discretization,
time-steps choice), and if / how the library can be extended to fit more
specific needs.

Similarly, there is no in-between the very short overview and trivial
example and the full API documentation (which is welcome but hard to decipher).

Some doc-strings lack of details : the allowed boundary conditions are well
described in the raised ValueError, but not in the doc-string itself.

There is more examples available on the repository than shown in the
documentation : using [Sphinx-Gallery](https://sphinx-gallery.github.io/stable
/index.html), these examples could easily be integrated in the documentation.

The software seems to embed a benchmark tool : it could be nice to have a
summary of the performance of implemented models with
and without numba and with and without parallelization.

## Community guidelines

It could be nice to add a Code of Conduct (see the [scipy one](https://docs
.scipy.org/doc/scipy/reference/dev/conduct/code_of_conduct.html) for an
extended example, and [here](https://www.contributor-covenant.org/) for a
standard one that you can easily adapt for your repository).
