### 2020-03-16 a306685
Added optional dependencies
