### 2020-03-16 23a1d33
Create CODE_OF_CONDUCT.md
Added code of conduct
