### 2020-03-16 2cbaf5b
Improved documentation
