+++ docs/source/quickstart/advanced_usage.rst
@@ -1,0 +1,321 @@
+Advanced usage
+^^^^^^^^^^^^^^
+
+Advanced boundary conditions
+""""""""""""""""""""""""""""
+Boundary conditions can be specified for both sides of each axis individually.
+For instance, we can enforce the value of a field to be `4` at the lower side
+and its derivative (in the outward direction) to be `2` on the upper side:
+
+.. code-block:: python
+
+    bc_lower = {'type': 'value', 'value': 4}
+    bc_upper = {'type': 'derivative', 'value': 2}
+    bc = [bc_lower, bc_upper]
+    
+    grid = pde.UnitGrid([16])
+    field = pde.ScalarField(grid)
+    field.laplace(bc)
+    
+
+Here, the laplace operator applied to the field in the last line will respect
+the boundary conditions.
+Note that it suffices to give the condition once if it is the same on both
+sides.
+For instance, to enforce a value of `3` on both side, one could simply specify
+:code:`bc = {'type': 'value', 'value': 3}`.
+
+
+Custom PDEs
+"""""""""""
+The `py-pde` package provides classes that help with implementing and analyzing
+partial differential equations.
+To implement a new PDE in a way that all of the machinery of `py-pde` can be
+used, one needs to subclass :class:`pde.pdes.base.PDEBase` and overwrite at 
+least the :meth:`pde.pdes.base.PDEBase.evolution_rate`.
+
+As an example we implement the `Kuramoto–Sivashinsky equation 
+<https://en.wikipedia.org/wiki/Kuramoto–Sivashinsky_equation>`_, given by    
+:math:`\partial_t u = - \nabla^4 u - \nabla^2 u - \frac{1}{2} |\nabla u|^2`,
+which describes the time evolution of a scalar field :math:`u`.
+A simple implementation of this equation could read 
+
+.. code-block:: python
+
+    class KuramotoSivashinskyPDE(PDEBase):
+        
+        def evolution_rate(self, state, t=0):
+            """ numpy implementation of the evolution equation """
+            state_lapacian = state.laplace(bc='natural')
+            return (- state_lapacian.laplace(bc='natural')
+                    - state_lapacian
+                    - 0.5 * state.gradient(bc='natural').to_scalar('squared_sum'))
+
+A slightly more advanced example would allow for class attributes that for
+instance define the boundary conditions and the diffusivity:
+
+.. code-block:: python
+
+    class KuramotoSivashinskyPDE(PDEBase):
+        
+        def __init__(self, diffusivity=1, bc='natural', bc_laplace='natural'):
+            """ initialize the class with a diffusivity and boundary conditions
+            for the actual field and its second derivative """
+            self.diffusivity = diffusivity
+            self.bc = bc
+            self.bc_laplace = bc_laplace
+        
+        def evolution_rate(self, state, t=0):
+            """ numpy implementation of the evolution equation """
+            state_lapacian = state.laplace(bc=self.bc)
+            state_gradient = state.gradient(bc=self.bc)
+            return (- state_lapacian.laplace(bc=self.bc_laplace)
+                    - state_lapacian
+                    - 0.5 * self.diffusivity * (state_gradient @ state_gradient))
+
+We here replaced the call to :code:`to_scalar('squared_sum')` by a 
+dot product with itself (using the `@` notation), which is equivalent.
+Note that the numpy implementation of the right hand side of the PDE is rather
+slow since it runs mostly in pure python and constructs a lot of intermediate
+field classes.
+While such an implementation is helpful for testing initial ideas, actual
+computations should be performed with compiled PDEs as described below.
+
+
+Operators without field classes
+"""""""""""""""""""""""""""""""
+This section explains how to use the low-level version of the field operators.
[…]
