### 2020-03-16 7401bc9
Fixed problem with included images in latex
