### 2020-03-18 8dcd3ef
Added code for solving Poisson's problem
Also added VectorField.from_expression constructor
Minor other improvements
