### 2020-03-18 69e4bed
Reorganized examples and added docstrings to them
