+++ examples/cartesian_grid.py
@@ -1,4 +1,9 @@
-#!/usr/bin/env python3
+"""
+Diffusion on a Cartesian grid
+=============================
+
+This example shows how to solve the diffusion equation on a Cartesian grid
+"""
 
 from pde import CartesianGrid, ScalarField, DiffusionPDE
 
