+++ examples/stochastic_simulation.py
@@ -1,13 +1,18 @@
-#!/usr/bin/env python3
+"""
+Stochastic simulation
+=====================
+
+This example illustrates how a stochastic simulation can be done.
+"""
 
 from pde import (KPZInterfacePDE, UnitGrid, ScalarField, MemoryStorage,
                  plot_kymograph)
 
 
-eq = KPZInterfacePDE(noise=1)              # define the SDE
 grid = UnitGrid([32])                      # generate grid
 state = ScalarField.random_harmonic(grid)  # generate initial condition
 
+eq = KPZInterfacePDE(noise=1)              # define the SDE
 storage = MemoryStorage()
 eq.solve(state, t_range=10, dt=0.01, tracker=storage.tracker(1))
-plot_kymograph(storage, show=True)
+plot_kymograph(storage)
