+++ examples/custom_pde_1d.py
@@ -1,4 +1,9 @@
-#!/usr/bin/env python3
+"""
+Custom PDE in one dimension
+===========================
+
+This example implements a PDE that is only defined in one dimension.
+"""
 
 from math import pi
 
@@ -22,11 +27,11 @@ class KortewegDeVriesPDE(PDEBase):
 # initialize the equation and the space
 grid = CartesianGrid([[0, 2*pi]], [32], periodic=True)
 state = ScalarField.from_expression(grid, "sin(x)")
-eq = KortewegDeVriesPDE()             
 
 # solve the equation and store the trajectory
 storage = MemoryStorage()
-eq.solve(state, t_range=3, tracker=['progress', storage.tracker(.1)])
+eq = KortewegDeVriesPDE()             
+eq.solve(state, t_range=3, tracker=storage.tracker(.1))
 
 # plot the trajectory as a space-time plot
-plot_kymograph(storage, show=True)
+plot_kymograph(storage)
