+++ examples/custom_pde.py
@@ -1,4 +1,9 @@
-#!/usr/bin/env python3
+"""
+Custom scalar PDE
+=================
+
+This example implements a custom scalar PDE.
+"""
 
 from pde import UnitGrid, ScalarField, PDEBase
 
@@ -17,9 +22,9 @@ class KuramotoSivashinskyPDE(PDEBase):
         return -state_grad.to_scalar('squared_sum') / 2 - state_lap - state_lap2 
 
 
-eq = KuramotoSivashinskyPDE()             # define the pde
 grid = UnitGrid([16, 16])                 # generate grid
 state = ScalarField.random_uniform(grid)  # generate initial condition
 
+eq = KuramotoSivashinskyPDE()             # define the pde
 result = eq.solve(state, t_range=10, dt=0.01)
 result.plot(show=True)
