### 2020-03-18 b43c9f6
Improved documentation
