+++ examples/README.txt
@@ -1,0 +1,5 @@
+Examples gallery
+================
+
+Below is a gallery of examples of the `py-pde` package, which illustrates some
+of the most important features of the package.
