+++ docs/source/manual/index.rst
@@ -1,13 +1,11 @@
-Getting started
-===============
+User manual
+===========
 
 
 .. toctree::
    :maxdepth: 3
     
    
-   overview
-   examples
    mathematical_basics
    advanced_usage
    contributing
