+++ docs/requirements.txt
@@ -1,0 +1,8 @@
+matplotlib>=3.1.0
+numpy>=1.18.0
+numba>=0.48.0
+scipy>=1.4.0
+sympy>=1.5.0
+sphinx-autodoc-annotation>=1.0
+sphinx-gallery>=0.5
+sphinx-rtd-theme>=0.4
