### 2020-03-18 84ef3d6
Remove movie example from gallery
