+++ docs/source/conf.py
@@ -253,6 +253,7 @@ sphinx_gallery_conf = {
      'examples_dirs': '../../examples',
      'gallery_dirs': 'examples_gallery',
      'filename_pattern': '/',
+     'ignore_pattern': r'.*movie.*',
      'capture_repr': (),
 }
 
