### 2020-03-18 f27c1d2
Added two examples
Improved layout of examples gallery
