+++ docs/source/conf.py
@@ -38,8 +38,8 @@ release = pde.__version__
 
 
 # -- Pre-processing ---------------------------------------------------
-from parse_examples import main
-main()
+# from parse_examples import main
+# main()
 
 
 rst_prolog = """
@@ -141,6 +141,10 @@ html_static_path = ['_static']
 #
 # html_sidebars = {}
 
+
+html_css_files = [
+    'custom.css',
+]
 
 # -- Options for HTMLHelp output ---------------------------------------------
 
