### 2020-03-18 b9e32f9
Add check for consistency of the numba-compiled implementations
