### 2020-03-18 f8f141d
Added matrix Laplacian to performance test
Improved performance by using proper sparse dot product
