### 2020-03-19 5deb186
Improved examples and documentation
