+++ examples/cartesian_grid.py
@@ -7,9 +7,10 @@ This example shows how to solve the diffusion equation on a Cartesian grid
 
 from pde import CartesianGrid, ScalarField, DiffusionPDE
 
-grid = CartesianGrid([[0, 1], [0, 2]], [3, 8])     # generate grid
-state = ScalarField.random_uniform(grid, 0.2, 0.3) # generate initial condition
+grid = CartesianGrid([[-1, 1], [0, 2]], [15, 8])  # generate grid
+state = ScalarField(grid)  # generate initial condition
+state.add_interpolated([0, 1], 1)
 
-eq = DiffusionPDE()                               # define the pde
-result = eq.solve(state, t_range=10, dt=0.005)    # solve it
+eq = DiffusionPDE()  # define the pde
+result = eq.solve(state, t_range=1, dt=0.005)
 result.plot(show=True)
