+++ examples/boundary_conditions.py
@@ -1,6 +1,6 @@
 """
-Boundary conditions
-===================
+Setting boundary conditions
+===========================
 
 This example shows how different boundary conditions can be specified.
 """
@@ -11,8 +11,8 @@ grid = UnitGrid([16, 16], periodic=[False, True])   # generate grid
 state = ScalarField.random_uniform(grid, 0.2, 0.3)  # generate initial condition
 
 # set boundary conditions `bc` for all axes
-bc_x_left = {'type': 'derivative', 'value': 0.1}  
-bc_x_right = {'type': 'value', 'value': 0}
+bc_x_left = {'derivative': 0.1}  
+bc_x_right = {'value': 'sin(y)'}
 bc_x = [bc_x_left, bc_x_right]
 bc_y = 'periodic' 
 eq = DiffusionPDE(bc=[bc_x, bc_y])
