### 2020-03-19 a367ac7
adjusted tolerance in test of Poisson solver
