+++ pde/grids/operators/tests/test_cartesian.py
@@ -320,7 +320,8 @@ def test_poisson_solver_general(dim):
     
     d = np.random.random(bcs.grid.shape)
     d -= d.mean()  # balance the right hand side
-    np.testing.assert_allclose(laplace(poisson(d)), d, err_msg=f'bcs = {bcs}')
+    np.testing.assert_allclose(laplace(poisson(d)), d, rtol=1e-5,
+                               err_msg=f'bcs = {bcs}')
 
 
 
