### 2020-03-19 86d0f93
Fixed a but in interpolation in 1d grids
Improved test coverage
