+++ pde/grids/tests/test_cylindrical.py
@@ -85,3 +85,14 @@ def test_setting_domain_cylindrical():
         grid.get_boundary_conditions(['no-flux', 'no-flux'])
         
         
+
+@pytest.mark.parametrize('periodic', [True, False])
+def test_polar_conversion(periodic):
+    """ test conversion to polar coordinates """
+    grid = CylindricalGrid(1, [-1, 1], [5, 5], periodic_z=periodic)
+    dists = grid.polar_coordinates_real([0, 0, 0])
+    assert np.all(0.09 <= dists) 
+    assert np.any(dists < 0.11)
+    assert np.all(dists <= np.sqrt(2)) 
+    assert np.any(dists > 0.8 * np.sqrt(2)) 
+        
