### 2020-03-20 05cf555
Removed old interpolate_to_grid method
