+++ pde/grids/operators/tests/test_spherical.py
@@ -70,21 +70,22 @@ def test_small_annulus(make_op, field):
 
 
 
+
 def test_grid_laplace():
-    """ test the spherical implementation of the laplace operator """
-    grid_sph = SphericalGrid(10, 11)
+    """ test the polar implementation of the laplace operator """
+    grid_sph = SphericalGrid(9, 11)
     grid_cart = CartesianGrid([[-5, 5], [-5, 5], [-5, 5]], [12, 10, 11]) 
      
-    a_1d = np.cos(grid_sph.axes_coords[0])
-    a_3d = grid_sph.interpolate_to_cartesian(a_1d, grid=grid_cart)
+    a_1d = ScalarField.from_expression(grid_sph, 'cos(r)')
+    a_3d = a_1d.interpolate_to_grid(grid_cart)
 
-    b_3d = grid_cart.get_operator('laplace', 'no-flux')(a_3d)
-    b_1d = grid_sph.get_operator('laplace', 'no-flux')(a_1d)
-    b_1d_3 = grid_sph.interpolate_to_cartesian(b_1d, grid=grid_cart)
+    b_3d = a_3d.laplace('natural')
+    b_1d = a_1d.laplace('natural')
+    b_1d_3 = b_1d.interpolate_to_grid(grid_cart)
      
     i = slice(1, -1)  # do not compare boundary points
-    np.testing.assert_allclose(b_1d_3[i, i, i], b_3d[i, i, i], rtol=0.2,
-                               atol=0.2)
+    np.testing.assert_allclose(b_1d_3.data[i, i, i], b_3d.data[i, i, i],
+                               rtol=0.2, atol=0.2)
      
      
 

[…]
