### 2020-03-21 4a3396a
Added automatic docstring replacements
Improved documentation
