+++ docs/source/conf.py
@@ -42,11 +42,6 @@ release = pde.__version__
 # main()
 
 
-rst_prolog = """
-.. include:: ../doc_replacements.txt
-"""
-
-
 # -- General configuration ---------------------------------------------------
 
 # If your documentation needs a minimal Sphinx version, state it here.
