### 2020-03-21 3214012
Improved docstrings further
