+++ pde/pdes/swift_hohenberg.py
@@ -13,6 +13,7 @@ from .base import PDEBase
 from ..fields import ScalarField
 from ..grids.boundaries.axes import BoundariesData
 from ..tools.numba import nb, jit
+from ..tools.docstrings import fill_in_docstring
 
 
         
@@ -31,6 +32,7 @@ class SwiftHohenbergPDE(PDEBase):
     explicit_time_dependence = False
 
 
+    @fill_in_docstring
     def __init__(self,
                  rate: float = .1,
                  kc2: float = 1.,
@@ -46,14 +48,12 @@ class SwiftHohenbergPDE(PDEBase):
             delta (float):
                 Parameter :math:`\delta` of the non-linearity
             bc:
-                The boundary conditions applied to the scalar field :math:`c`.
-                The default value ('natural') imposes periodic boundary
-                conditions for axes in which the grid is periodic and vanishing
-                derivatives for all other axes. Alternatively, specific boundary
-                conditions can be set for all axes individually. 
+                {ARG_BOUNDARIES}
             bc_lap:
                 The boundary conditions applied to the second derivative of the
-                scalar field :math:`c`. If `None`, the same 
+                scalar field :math:`c`. If `None`, the same boundary condition
+                as `bc` is chosen. Otherwise, this supports the same options as
+                `bc`.
         """
         super().__init__()
         
