+++ pde/pdes/wave.py
@@ -13,6 +13,7 @@ from .base import PDEBase
 from ..fields import ScalarField, FieldCollection
 from ..grids.boundaries.axes import BoundariesData
 from ..tools.numba import nb, jit
+from ..tools.docstrings import fill_in_docstring
 
 
         
@@ -35,6 +36,7 @@ class WavePDE(PDEBase):
     explicit_time_dependence = False
     
 
+    @fill_in_docstring
     def __init__(self, speed: float = 1,
                  bc: BoundariesData = 'natural'):
         """ 
@@ -42,11 +44,7 @@ class WavePDE(PDEBase):
             speed (float):
                 The speed :math:`c` of the wave
             bc:
-                The boundary conditions applied to the scalar field :math:`c`.
-                The default value ('natural') imposes periodic boundary
-                conditions for axes in which the grid is periodic and vanishing
-                derivatives for all other axes. Alternatively, specific boundary
-                conditions can be set for all axes individually. 
+                {ARG_BOUNDARIES}
         """
         super().__init__()
         
