### 2020-03-21 704d848
Added lgtm config
Removed some problems spotted by lgtm
