+++ pde/fields/base.py
@@ -10,8 +10,8 @@ import operator
 import logging
 from pathlib import Path 
 from abc import ABCMeta, abstractmethod, abstractproperty
-from typing import (Tuple, Callable, Optional, Union, Any, Dict,
-                    List, TypeVar, Iterator, TYPE_CHECKING)  # @UnusedImport
+from typing import (Tuple, Callable, Optional, Union, Any, Dict, TypeVar, 
+                    TYPE_CHECKING)
 
 import numpy as np
 from scipy import interpolate, ndimage
