+++ pde/pdes/kpz_interface.py
@@ -4,7 +4,7 @@ The Kardar–Parisi–Zhang (KPZ) equation describing the evolution of an interf
 .. codeauthor:: David Zwicker <david.zwicker@ds.mpg.de> 
 """
 
-from typing import (Callable, Dict, Any, Counter, List)  # @UnusedImport
+from typing import Callable
 
 import numpy as np
 
