### 2020-03-22 5069571
Added `project` and `slice` methods to scalar fields
Refactored integration and how cell volumes are stored
