### 2020-03-23 bfe793a
Added docstring to ScalarField.slice method
