### 2020-03-23 aef5765
Multiple smaller improvements
