### 2020-03-23 9be7fd7
Improved documentation of boundary conditions
