### 2020-03-23 d735a6c
Added example for showing data manipulations
