+++ TODO.md
@@ -1,10 +1,5 @@
 TODO
 ====
-* Add example for dataanalysis uysing project, slice, fluctuations, average, ...
-    - also display plotting multiple things
-* Use project and slice in plotting
-* Give examples for boundary conditions in the boundaries module description
-    - link to this from the normal arguments
 * Think about introducing data class that holds integrated, global variables
 	- this might be helpful to implement lagrange multipliers and the like
 * Fix progress bar when starting from non-zero t_start?
