### 2020-03-24 433ffec
Adjusted signature of method plot_vector in base class
