### 2020-03-24 79cc6a5
Adjusted lgtm configuration
