### 2020-03-24 1b77fed
Added special positions in slicing
