+++ pde/solvers/tests/test_explicit.py
@@ -18,15 +18,15 @@ def test_compare_explicit():
     field = ScalarField.random_uniform(grid, -1, 1)
     eq = DiffusionPDE()
     
-    c1 = Controller(ExplicitSolver(eq), t_range=1, tracker=None)
-    s1 = c1.run(field, dt=1e-3)
+    c1 = Controller(ExplicitSolver(eq), t_range=.1, tracker=None)
+    s1 = c1.run(field, dt=2e-3)
     
-    c2 = Controller(ExplicitSolver(eq, scheme='runge-kutta'), t_range=1,
+    c2 = Controller(ExplicitSolver(eq, scheme='runge-kutta'), t_range=.1,
                     tracker=None)
     with np.errstate(under='ignore'):
-        s2 = c2.run(field, dt=1e-3)
+        s2 = c2.run(field, dt=2e-3)
                                
-    np.testing.assert_allclose(s1.data, s2.data, rtol=1e-3, atol=1e-3)
+    np.testing.assert_allclose(s1.data, s2.data, rtol=1e-2, atol=1e-2)
 
 
 
