+++ pde/solvers/tests/test_generic.py
@@ -3,6 +3,7 @@
 '''
 
 import numpy as np
+import pytest
 
 from ...fields import ScalarField
 from ...grids import UnitGrid
@@ -28,20 +29,21 @@ def test_solver_in_pde_class():
 
 
 
-def test_compare_solvers():
+@pytest.mark.parametrize('solver_class', [ExplicitSolver, ImplicitSolver,
+                                          ScipySolver])
+def test_compare_solvers(solver_class):
     """ compare several solvers """
-    field = ScalarField.random_uniform(UnitGrid([16, 16]), -1, 1)
+    field = ScalarField.random_uniform(UnitGrid([8, 8]), -1, 1)
     eq = DiffusionPDE()
     
     # ground truth
-    c1 = Controller(ExplicitSolver(eq, scheme='runge-kutta'), t_range=1,
+    c1 = Controller(ExplicitSolver(eq, scheme='runge-kutta'), t_range=.1,
                     tracker=None)
-    s1 = c1.run(field, dt=1e-3)
+    s1 = c1.run(field, dt=5e-3)
     
-    for solver_class in [ExplicitSolver, ImplicitSolver, ScipySolver]:
-        c2 = Controller(solver_class(eq), t_range=1, tracker=None)
-        with np.errstate(under='ignore'):
-            s2 = c2.run(field, dt=1e-3)
-                               
-        np.testing.assert_allclose(s1.data, s2.data, rtol=1e-3, atol=1e-3)
+    c2 = Controller(solver_class(eq), t_range=.1, tracker=None)
+    with np.errstate(under='ignore'):
+        s2 = c2.run(field, dt=5e-3)
+                           
+    np.testing.assert_allclose(s1.data, s2.data, rtol=1e-2, atol=1e-2)
 
