### 2020-03-25 afbabaf
Update README.md
Added code quality to readme
