### 2020-03-25 1aa7888
Suppress false positive lgtm notice
