### 2020-03-25 cfb8bfc
Improved typing
