### 2020-03-25 932577e
Improved testing and documentation
