+++ pde/visualization/plotting.py
@@ -14,7 +14,7 @@ Functions and classes for plotting simulation data
 
 import logging
 import warnings
-from typing import Union, Callable, Optional, Any, List  # @UnusedImport
+from typing import (Union, Callable, Optional, Any, Dict, List, Tuple)
 
 import numpy as np
 
@@ -23,6 +23,9 @@ from ..fields.base import FieldBase, DataFieldBase
 from ..storage.base import StorageBase
 from ..tools.misc import display_progress
 from ..tools.docstrings import fill_in_docstring
+
+
+ColorScaleData = Union[str, float, Tuple[float, float]]
 
 
 
@@ -142,13 +145,18 @@ class ScalarFieldPlot():
     """ class managing compound plots of scalar fields """
     
     @fill_in_docstring
-    def __init__(self, fields: FieldBase, quantities=None, show: bool = True):
+    def __init__(self, fields: FieldBase,
+                 quantities=None,
+                 color_scale: ColorScaleData = 'automatic',
+                 show: bool = True):
         """
         Args:
             fields:
                 Collection of fields
             quantities:
                 {ARG_PLOT_QUANTITIES}
+            color_scale (str, float, tuple of float):
+                {ARG_COLOR_SCALE}
             show (bool):
                 Flag determining whether to show a plot. If `False`, the plot is
                 kept in the background, which can be useful if it only needs to
@@ -159,7 +167,8 @@ class ScalarFieldPlot():
         
         self.grid = fields.grid
         example_image = fields.get_image_data()
-        self.quantities = self._prepare_quantities(fields, quantities)
+        self.quantities = self._prepare_quantities(fields, quantities,
+                                                   color_scale=color_scale)
         self.show = show
              
         num_rows = len(self.quantities)
@@ -179,16 +188,25 @@ class ScalarFieldPlot():
             img_row = []
             for j, panel in enumerate(panel_row):
                 # determine scale of the panel
-                scale = panel.get('scale', 1)
-                try:
-                    vmin, vmax = scale
-                except TypeError:
-                    vmin, vmax = 0, scale
+                scale = panel.get('scale', color_scale)
+                if scale == 'automatic':
+                    vmin, vmax = None, None
+                elif scale == 'unity':
+                    vmin, vmax = 0, 1
+                elif scale == 'symmetric':
+                    vmin, vmax = -1, 1
+                else:
+                    try:
+                        vmin, vmax = scale
+                    except TypeError:
+                        vmin, vmax = 0, scale
                     
                 # determine colormap
                 cmap = panel.get('cmap')
                 if cmap is None:
-                    if np.isclose(-vmin, vmax):
+                    if vmin is None or vmax is None:
+                        cmap = cm.PiYG
+                    elif np.isclose(-vmin, vmax):
                         cmap = cm.coolwarm
                     elif np.isclose(vmin, 0):
                         cmap = cm.gray
@@ -204,16 +222,21 @@ class ScalarFieldPlot():
                 ax.set_title(panel.get('title', ''))
                 ax.set_xlabel(example_image['label_x'])
                 ax.set_ylabel(example_image['label_y'])
-#                 ax.axison = False
+                
+                # store the default value alongside
+                img._default_vmin_vmax = vmin, vmax
                 img_row.append(img)
             self.images.append(img_row)
             
         self.show_data(fields)    
         
     
-    @fill_in_docstring
     @classmethod
-    def from_storage(cls, storage: StorageBase, quantities=None):
+    @fill_in_docstring
+    def from_storage(cls, storage: StorageBase,
+                     quantities=None,
+                     color_scale: ColorScaleData = 'automatic') \
+            -> "ScalarFieldPlot":
[…]
