### 2020-03-25 4bc7597
Allow setting color scale in PlotTracker
