### 2020-03-25 71081ee
Remove unnecessary methods in Parameters
