### 2020-03-25 28ea0cc
Added new examples
Improved other examples
