+++ examples/plot_vector_field.py
@@ -1,8 +1,9 @@
-"""
+r"""
 Plotting a vector field
 =======================
 
-This example shows how to initialize and visualize a vector field.
+This example shows how to initialize and visualize the vector field 
+:math:`\boldsymbol u = \bigl(\sin(x), \cos(x)\bigr)`.
 """
 
 from pde import CartesianGrid, VectorField
