### 2020-03-25 01caeab
Added Brusselator to examples
