### 2020-03-27 2b76a92
Improved expressions module
Increased test coverage
