### 2020-03-27 dc7d1ba
Adjust codecov settings
