### 2020-03-27 c5e5733
Small improvements to parameters module
Improved test coverage
