### 2020-03-27 7cf171e
Improved code style
