### 2020-03-28 f989c45
Fixed some problems with equality operators
