### 2020-03-29 3f0da0b
Made some attributes of the field classes read-only
