### 2020-03-29 4f8345e
Made some properties of grids read-only
