### 2020-03-30 #2158 @david-zwicker
Dear @celliern, thank you very much for your suggestions! I managed to include most of your points in the revised code. Thank you especially for suggesting the example gallery – I didn't know about this, but I think it helps showcasing the features of the package tremendously.

I'm looking forward to the remaining comments, but I obviously understand if this will now take much longer than usual. Stay safe and healthy!
