### 2020-03-30 #2158 @celliern
@david-zwicker Nice your gallery is wonderful !

Two minor points :

- "Do the authors clearly state what problems the software is designed to solve and who the target audience is?"
This could be highlighted  in the "Getting started" : I guess that such software is designed with researcher and student in mind (more than engineer / industrial target) due to the well known finite difference limitation (especially on complex geometry).

- The CoC is in the repo, but should be linked in the doc and / or in the repository README.
I'm looking at the paper today.

As a side suggestion (not blocking for the publication) : you will have a strong limitation on hyperbolic equations. This can be (partially) mitigated by introducing "advection" operator that could either implement really simple Gudunov finite volume scheme or upwind finite difference scheme. This will extend the scope of the software.

Stay safe and healthy as well !
