### 2020-03-30 e708552
Add a test for interpolation of mutuable fields
