### 2020-03-30 #2158 @celliern
For the paper:
- there is still some typo : they could be easily spotted by a spell checker.
- `pyclaw` is the Clawpack python interface, clawpack is specialized in finite volume (especially Gudunov-like methods), not finite difference.
- you could cite the [Dedalus Project](http://dedalus-project.org/) for the pseudo-spectral methods.

Otherwise, the paper look fine. I will recommand the paper / software for publication once these different point are fixed.

You have written a nice scientic tool. Nice work !
