### 2020-03-30 d2eb0d1
Test for in-place smoothing of fields
