### 2020-03-30 4d87a0d
Allow explicit removing of plot title
