### 2020-03-30 7eed9c5
Improved documentation
