### 2020-03-30 37a7546
Fixed equality operators of fields
Added hints for lgtm
