+++ pde/fields/base.py
@@ -267,10 +267,10 @@ class FieldBase(metaclass=ABCMeta):
             
     
     def __eq__(self, other):
-        """ test for equality """
+        """ test fields for equality, ignoring the label """
         if not isinstance(other, self.__class__):
             return NotImplemented
-        return (self.state == other.state and
+        return (self.grid == other.grid and
                 np.array_equal(self.data, other.data))
     
 
