### 2020-03-30 1f14d13
Improved Poisson solver
Improved logging and messages of the Poisson solver
Improved some tests
