### 2020-03-30 2d7b779
Adjusted codecov configuration
