+++ codecov.yml
@@ -1,5 +1,8 @@
 coverage:
   status:
+    project:
+      default:
+        threshold: 1%
     patch: off
 ignore:
   - "**/tests/*"
