### 2020-03-31 70224c2
Fixed minor issues detected by lgtm
