### 2020-03-31 0a5a0f4
Added code of conduct to documentation
