### 2020-03-31 zwicker-group/py-pde#2 @mstimberg
Requirements for sphinx documentation
As part of openjournals/joss-reviews#2158 I found two minor issues with building the sphinx documentation:
* The requirements stated on https://py-pde.readthedocs.io/en/latest/getting_started.html#prerequisites do not include `sphinx-gallery`.
* The `docs/requirements.txt` file does not include `pillow`, but it seems to be necessary to generate thumbnails for the sphinx gallery.
