### 2020-03-31 zwicker-group/py-pde#3 @mstimberg
tqdm in jupyter notebook
[openjournals/joss-reviews#2158]

When I run the example jupyter notebooks, the tracker displaying an error bar with `tqdm` fails if `ipywidgets` is not installed.
