### 2020-03-31 zwicker-group/py-pde#4 @mstimberg
Suggestion: test dependencies
[openjournals/joss-reviews#2158]

For consistency, I'd suggest mentioning the dependencies for running tests already on the [Getting Started](https://py-pde.readthedocs.io/en/latest/getting_started.html#prerequisites) page of the documentation (in the same way as for the documentation dependencies).

Given that "nobody reads the documentation", you might also want to consider a check for the presence of `pytest` etc. directly in the shell scripts with a helpful error message for the user. For example, if the user does not have `pytest-xdist` installed, running `tests_parallel.sh` will currently fail with "unknown option -n".
