### 2020-03-31 zwicker-group/py-pde#5 @mstimberg
Suggestion: add CI for Windows/OS X
[openjournals/joss-reviews#2158]

The documentation states "The code is tested under Linux, Windows, and macOS", but the repository currently only runs the test suite on Linux via Travis. Travis also supports testing under OS X, and there are other (free for FOSS) services like https://www.appveyor.com/ or https://azure.microsoft.com/en-us/services/devops/pipelines/ that test across all three OS.
