### 2020-03-31 zwicker-group/py-pde#6 @mstimberg
Suggestion: example gallery for notebooks
[openjournals/joss-reviews#2158]

The example gallery is great, you might consider including the notebooks in `examples/jupyter` in the documentation as well. With the [`sphinx-nbexamples`](https://sphinx-nbexamples.readthedocs.io) extension, this should work very similar to the `sphinx-gallery` setup.
