### 2020-03-31 #2158 @mstimberg
I join @celliern in congratulating you to your software, this is really an excellent tool and I have rarely seen that much attention to code quality in scientific software (type annotations, etc.)!

I have reported two minor issues (zwicker-group/py-pde#2, zwicker-group/py-pde#3) and three suggestions (zwicker-group/py-pde#4, zwicker-group/py-pde#5, zwicker-group/py-pde#5) directly in the `py-pde` repository, but @celliern's thorough review certainly made my job much easier.

Fixing the typos in the article is the only real blocker for me – if I wanted to be really nitpicky, I could also mention that the capitalization in the references list is somewhat inconsistent (e.g. "Python" vs. "python").

Pending these minor issues I can easily recommend this paper for publication. :+1:
