### 2020-03-31 #2158 @danielskatz
> Fixing the typos in the article is the only real blocker for me – if I wanted to be really nitpicky, I could also mention that the capitalization in the references list is somewhat inconsistent (e.g. "Python" vs. "python"

Please do be nitpicky - we need to fix all of these before publication :)
