### 2020-03-31 zwicker-group/py-pde#2 @david-zwicker
Thank you for the report. The documentation was indeed outdated and did not list all packages required to build the documentation. However, I did not need to install `pillow` explicitly to build the documentation. Are you seeing a specific error? Note that I'm currently having trouble with `sphinx-gallery` due to a [bug in that package](https://github.com/sphinx-gallery/sphinx-gallery/issues/634).
