### 2020-03-31 9882229
Added requirements for tests and improved documentation on install
