### 2020-03-31 cbcf7c5
Use standard tqdm if ipywidgets is not available
