### 2020-03-31 zwicker-group/py-pde#3 @david-zwicker
Thank you for the report. I feel this is more a bug of `tqdm.auto`, but I added a simple safe-guard checking for `ipywidgets` to circumvent this.
