### 2020-03-31 zwicker-group/py-pde#4 @david-zwicker
Thank you for the suggestion. I added a `requirements.txt` to the `tests` directory, which hopefully helps with setting up a suitable test environment. I also now mention this file and the one in the `docs` folder in the documentation.
