### 2020-03-31 zwicker-group/py-pde#5 @david-zwicker
I use github Actions to run the tests under macOS, windows, and linux. I haven't created a badge for this, since I don't want too many badges in the readme.
