### 2020-03-31 zwicker-group/py-pde#2 @mstimberg
Earlier, I got the following error after calling `make html`:
```pytb
Exception occurred:
  File "/home/marcel/anaconda3/envs/pypde/lib/python3.7/site-packages/sphinx_gallery/utils.py", line 27, in _get_image
    % (exc,))
RuntimeError: Could not import pillow, which is required to rescale images (e.g., for thumbnails): No module named 'PIL'
```
However, I cannot reproduce the problem any more after uninstalling `pillow` or running it in a new environment without it :roll_eyes:  Please feel free to close the issue.
