### 2020-03-31 zwicker-group/py-pde#5 @mstimberg
Ah, great! Sorry I overlooked this.
