### 2020-03-31 zwicker-group/py-pde#2 @david-zwicker
This is indeed weird. However, if `sphinx-gallery` requires `pillow`, they should probably list it in their requirements. I certainly do not import it anywhere in my code directly.
