### 2020-03-31 zwicker-group/py-pde#2 @mstimberg
Oh, silly me. I *can* reproduce it when I delete the `docs/source/examples_gallery/images/` directory (I ran `make clean` but that only deletes the `docs/build` directory).

The `sphinx-gallery` [README file](https://github.com/sphinx-gallery/sphinx-gallery/blob/master/README.rst) states:

> Sphinx-Gallery has also support for packages like:
>
>     Matplotlib
>     Seaborn
>     Mayavi
>
> For much of this functionality, you will need pillow.

So I think for `py-pde`'s documentation it indeed seems to be a dependency.
