### 2020-03-31 a4b9929
Added Pillow to requirements for docs
