### 2020-03-31 zwicker-group/py-pde#2 @david-zwicker
Good catch! I added the package to `docs/requirements.txt`.
