### 2020-03-31 8d1c0f8
Improved landing package of documentation
Improved one example
