+++ examples/custom_pde_brusselator.py
@@ -40,10 +40,9 @@ class BrusselatorPDE(PDEBase):
         """ pure python implementation of the PDE """
         u, v = state
         rhs = state.copy()
-        rhs[0] = (self.diffusivity[0] * u.laplace(self.bc)
-                  + self.a - (self.b + 1) * u + u**2 * v)
-        rhs[1] = (self.diffusivity[1] * v.laplace(self.bc)
-                  + self.b * u - u**2 * v)
+        d0, d1 = self.diffusivity
+        rhs[0] = d0 * u.laplace(self.bc) + self.a - (self.b + 1) * u + u**2 * v
+        rhs[1] = d1 * v.laplace(self.bc) + self.b * u - u**2 * v
         return rhs
     
     def _make_pde_rhs_numba(self, state):
