+++ docs/source/index.rst
@@ -8,13 +8,21 @@ partial differential equations (PDEs) of the form
 	\partial_t u(\boldsymbol x, t) = \mathcal D[u(\boldsymbol x, t)] 
 		+ \eta(u, \boldsymbol x, t) \;,
 
-where :math:`\mathcal D` is a (non-lienar) differential operator that defines
+where :math:`\mathcal D` is a (non-linear) differential operator that defines
 the time evolution of a (set of) physical fields :math:`u` with possibly
 tensorial character, which depend on spatial coordinates :math:`\boldsymbol x`
 and time :math:`t`.
-The framework also supports stochastic differential equations, where the noise
-is represented by :math:`\eta` above. Note that we here represent stochastic
-differential equations in the Itô representation.
+The framework also supports stochastic differential equations in the Itô
+representation, where the noise is represented by :math:`\eta` above.
+
+The main audience for the package are researchers and students who want to
+investigate the behavior of a PDE and get an intuitive understanding of the
+role of the different terms and the boundary conditions.
+To support this, `py-pde` evaluates PDEs using the methods of lines with a
+finite-difference approximation of the differential operators.
+Consequently, the mathematical operator :math:`\mathcal D` can be naturally
+translated to a function evaluating the evolution rate of the PDE.
+
 
 
 **Contents**
