### 2020-03-31 5bf0e84
Fixed small things in the paper
