+++ docs/paper/paper.bib
@@ -16,6 +16,24 @@ title = {Numba: A LLVM-Based Python JIT Compiler}, year = {2015}, isbn = {978145
   journal = {Journal of Open Source Software}
 }
 
+
+@ARTICLE{dedalus,
+       author = {{Burns}, Keaton J. and {Vasil}, Geoffrey M. and {Oishi}, Jeffrey S. and
+         {Lecoanet}, Daniel and {Brown}, Benjamin P.},
+        title = "{Dedalus: A Flexible Framework for Numerical Simulations with Spectral Methods}",
+      journal = {arXiv e-prints},
+     keywords = {Astrophysics - Instrumentation and Methods for Astrophysics, Physics - Computational Physics, Physics - Fluid Dynamics},
+         year = 2019,
+        month = may,
+          eid = {arXiv:1905.10388},
+        pages = {arXiv:1905.10388},
+archivePrefix = {arXiv},
+       eprint = {1905.10388},
+ primaryClass = {astro-ph.IM},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2019arXiv190510388B},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
 @article{Fenics,
   title = {The FEniCS Project Version 1.5},
   author = {Martin S. Aln{\ae}s and Jan Blechta and Johan Hake and August Johansson and Benjamin Kehlet and Anders Logg and Chris Richardson and Johannes Ring and Marie E. Rognes and Garth N. Wells},
