### 2020-03-31 cb4d869
Improved paper
Updated readme
