+++ docs/paper/paper.bib
@@ -1,6 +1,8 @@
 @inproceedings{numba,
 author = {Lam, Siu Kwan and Pitrou, Antoine and Seibert, Stanley},
-title = {Numba: A LLVM-Based Python JIT Compiler}, year = {2015}, isbn = {9781450340052}, publisher = {Association for Computing Machinery}, address = {New York, NY, USA}, url = {https://doi.org/10.1145/2833157.2833162}, doi = {10.1145/2833157.2833162}, booktitle = {Proceedings of the Second Workshop on the LLVM Compiler Infrastructure in HPC}, articleno = {Article 7}, numpages = {6}, keywords = {compiler, LLVM, Python}, location = {Austin, Texas}, series = {LLVM ’15} }
+title = {{Numba: A LLVM-Based Python JIT Compiler}},
+year = {2015},
+isbn = {9781450340052}, publisher = {Association for Computing Machinery}, address = {New York, NY, USA}, url = {https://doi.org/10.1145/2833157.2833162}, doi = {10.1145/2833157.2833162}, booktitle = {Proceedings of the Second Workshop on the LLVM Compiler Infrastructure in HPC}, articleno = {Article 7}, numpages = {6}, keywords = {compiler, LLVM, Python}, location = {Austin, Texas}, series = {LLVM ’15} }
 
 
 @article{Cellier2019,
@@ -12,7 +14,7 @@ title = {Numba: A LLVM-Based Python JIT Compiler}, year = {2015}, isbn = {978145
   number = {38},
   pages = {1356},
   author = {Nicolas Cellier and Christian Ruyer-Quil},
-  title = {scikit-finite-diff, a new tool for PDE solving},
+  title = {{scikit-finite-diff, a new tool for PDE solving}},
   journal = {Journal of Open Source Software}
 }
 
@@ -35,7 +37,7 @@ archivePrefix = {arXiv},
 }
 
 @article{Fenics,
-  title = {The FEniCS Project Version 1.5},
+  title = {{The FEniCS Project Version 1.5}},
   author = {Martin S. Aln{\ae}s and Jan Blechta and Johan Hake and August Johansson and Benjamin Kehlet and Anders Logg and Chris Richardson and Johannes Ring and Marie E. Rognes and Garth N. Wells},
   year = {2015},
   journal = {Archive of Numerical Software},
@@ -71,7 +73,7 @@ archivePrefix = {arXiv},
 }
 
 @article{SfePy,
-  title =        {Multiscale finite element calculations in Python using SfePy},
+  title =        {{Multiscale finite element calculations in Python using SfePy}},
   ISSN =         {1572-9044},
   url =          {https://doi.org/10.1007/s10444-019-09666-0},
   DOI =          {10.1007/s10444-019-09666-0},
@@ -127,7 +129,7 @@ archivePrefix = {arXiv},
 
 
 @article{sympy,
- title = {SymPy: symbolic computing in Python},
+ title = {{SymPy: symbolic computing in Python}},
[…]
