### 2020-04-01 7c5937c
Added support for writing results of DataTrackers to files
