### 2020-04-01 4184c91
Updated examples to hopefully fix a problem in sphinx-gallery
