+++ examples/plot_cylindrical_field.py
@@ -10,4 +10,4 @@ from pde import CylindricalGrid, ScalarField
 
 grid = CylindricalGrid(radius=3, bounds_z=[0, 4], shape=16)
 field = ScalarField.from_expression(grid, 'sqrt(z) * exp(-r**2)')
-field.plot(title='Scalar field in cylindrical coordinates')
+field.plot(title='Scalar field in cylindrical coordinates', show=True)
