+++ examples/custom_pde_sir.py
@@ -3,9 +3,9 @@ Custom PDE: SIR model
 =====================
 
 This example implements a `spatially coupled SIR model 
-<https://en.wikipedia.org/wiki/Compartmental_models_in_epidemiology>`_, which
-defines the following dynamics for the density of susceptible, infected, and
-recovered individuals.
+<https://en.wikipedia.org/wiki/Compartmental_models_in_epidemiology>`_ with 
+the following dynamics for the density of susceptible, infected, and recovered
+individuals:
 
 .. math::
 
