+++ examples/custom_pde.py
@@ -25,9 +25,9 @@ class KuramotoSivashinskyPDE(PDEBase):
         return -state_grad.to_scalar('squared_sum') / 2 - state_lap - state_lap2 
 
 
-grid = UnitGrid([16, 16])                 # generate grid
+grid = UnitGrid([16, 16])  # generate grid
 state = ScalarField.random_uniform(grid)  # generate initial condition
 
-eq = KuramotoSivashinskyPDE()             # define the pde
+eq = KuramotoSivashinskyPDE()  # define the pde
 result = eq.solve(state, t_range=10, dt=0.01)
 result.plot(show=True)
