### 2020-04-01 #2158 @david-zwicker
Thank you very much for your constructive and helpful comments. I think I implemented almost all of them, except for the following three:

* **Adding the benchmarks to the documentations**: I'm not quite sure how to do this, so that it is useful. In particular, the performance will depend quite heavily on the hardware on which the code runs. Moreover, the results of the benchmark need to be cached somehow, so they are not re-run every time the documentation is updated. However, I also do not want to include results obtained on my local machine in the repository. Taken together, I'm not quite sure how to implement this so it's useful.
* **Adding notebook-gallery**: I had some issues with the normal gallery and thus would not like to increase the complexity of this solution. Moreover, there is a binder link in the Readme that people can use to directly play with the code. This is arguably more useful than static notebooks.
* **Advection operators**: This would certainly be useful, but I'm not very experienced with this and I'm afraid that a general enough solution is not trivial (even restricted to Cartesian grids) – All examples I found were for 1d systems with a pre-defined advection direction.

However, all your other comments were extremely useful and made the project and in particular the documentation much better. I also revised the paper (I'm sorry for all the spelling mistakes). Let me know if you have further suggestions for improvement!
