### 2020-04-01 #2158 @celliern
@david-zwicker No problem, there was non blocking suggestion ;)

For the advection specialized schemes I will try to do a pull request when I will have the time (I hope soon?).

@xuanxu I also had a final look on the documentation and paper, and I ecommend the paper publication !
