### 2020-04-03 884e00e
fix SymPy's doi
