+++ docs/paper/paper.bib
@@ -159,12 +159,12 @@ archivePrefix = {arXiv},
 	Author = {{Virtanen}, Pauli and {Gommers}, Ralf and {Oliphant}, Travis E. and {Haberland}, Matt and {Reddy}, Tyler and {Cournapeau}, David and {Burovski}, Evgeni and {Peterson}, Pearu and {Weckesser}, Warren and {Bright}, Jonathan and {van der Walt}, St{\'e}fan J. and {Brett}, Matthew and {Wilson}, Joshua and {Jarrod Millman}, K. and {Mayorov}, Nikolay and {Nelson}, Andrew R.~J. and {Jones}, Eric and {Kern}, Robert and {Larson}, Eric and {Carey}, CJ and {Polat}, {\.I}lhan and {Feng}, Yu and {Moore}, Eric W. and {Vand erPlas}, Jake and {Laxalde}, Denis and {Perktold}, Josef and {Cimrman}, Robert and {Henriksen}, Ian and {Quintero}, E.~A. and {Harris}, Charles R and {Archibald}, Anne M. and {Ribeiro}, Ant{\^o}nio H. and {Pedregosa}, Fabian and {van Mulbregt}, Paul and {Contributors}, SciPy 1. 0},
 	Date-Added = {2020-02-20 16:09:25 +0100},
 	Date-Modified = {2020-02-20 16:09:45 +0100},
-	Doi = {https://doi.org/10.1038/s41592-019-0686-2},
+	Doi = {10.1038/s41592-019-0686-2},
 	Journal = {Nature Methods},
 	Title = {{SciPy 1.0: Fundamental Algorithms for Scientific Computing in Python}},
 	Year = {2020}}
 
-	
+
 @article{tqdm,
   doi = {10.21105/joss.01277},
   url = {https://doi.org/10.21105/joss.01277},
@@ -177,4 +177,3 @@ archivePrefix = {arXiv},
   title = {{`tqdm`: A Fast, Extensible Progress Meter for Python and CLI}},
   journal = {Journal of Open Source Software}
 }
-	
