+++ docs/paper/paper.bib
@@ -56,7 +56,9 @@ archivePrefix = {arXiv},
         Pages = {C210--C231},
         Title = {{PyClaw: Accessible, Extensible, Scalable Tools for Wave Propagation Problems}},
         Volume = {34},
-        Year = {2012}}
+        Year = {2012},
+        doi = {10.1137/110856976}
+}
 
 
 @article{FiPy,
