### 2020-04-03 #2158 @david-zwicker
I just published the latest package on github and Zenodo:

The GitHub version is v0.5 and the DOI is https://doi.org/10.5281/zenodo.3739300

> On Apr 3, 2020, at 14:32, Juanjo Bazán <notifications@github.com> wrote:
>
>
> OK @david-zwicker, everything looks good, here are the next steps:
>
> 	• Please release a new tagged version from the current master so it includes all the changes made during the review process
> 	• Then archive that latest release in Zenodo
> 	• Check the Zenodo deposit has the correct metadata: title and author name should match the paper; you may also add your ORCID.
> Once you do that please report here the version number and archive DOI
>
