+++ docs/paper/paper.bib
@@ -1,8 +1,19 @@
 @inproceedings{numba,
-author = {Lam, Siu Kwan and Pitrou, Antoine and Seibert, Stanley},
-title = {{Numba: A LLVM-Based Python JIT Compiler}},
-year = {2015},
-isbn = {9781450340052}, publisher = {Association for Computing Machinery}, address = {New York, NY, USA}, url = {https://doi.org/10.1145/2833157.2833162}, doi = {10.1145/2833157.2833162}, booktitle = {Proceedings of the Second Workshop on the LLVM Compiler Infrastructure in HPC}, articleno = {Article 7}, numpages = {6}, keywords = {compiler, LLVM, Python}, location = {Austin, Texas}, series = {LLVM ’15} }
+  author = {Lam, Siu Kwan and Pitrou, Antoine and Seibert, Stanley},
+  title = {{Numba: A LLVM-Based Python JIT Compiler}},
+  year = {2015},
+  isbn = {9781450340052},
+  publisher = {Association for Computing Machinery},
+  address = {New York, NY, USA},
+  url = {https://doi.org/10.1145/2833157.2833162},
+  doi = {10.1145/2833157.2833162},
+  booktitle = {Proceedings of the Second Workshop on the {LLVM} Compiler Infrastructure in {HPC}},
+  articleno = {Article 7},
+  numpages = {6},
+  keywords = {compiler, LLVM, Python},
+  location = {Austin, Texas},
+  series = {LLVM ’15},
+}
 
 
 @article{Cellier2019,
@@ -101,7 +112,7 @@ archivePrefix = {arXiv},
 @Article{ipython,
   Author    = {P\'erez, Fernando and Granger, Brian E.},
   Title     = {{IP}ython: a System for Interactive Scientific Computing},
-  Journal   = {Computing in Science and Engineering},
+  Journal   = {Computing in Science \& Engineering},
   Volume    = {9},
   Number    = {3},
   Pages     = {21--29},
