### 2020-04-03 #2158 @danielskatz
👋 @david-zwicker - I've suggested some minor changes in https://github.com/zwicker-group/py-pde/pull/8/files
