### 2020-04-03 #2158 @david-zwicker
Thank you everyone for the very smooth process and for your numerous suggestions that surely improved the quality of the code, the documentation, and the paper.

There is some small issue still: I cannot see the final pdf of the paper and I actually receive a 404 when I want to download it. Is this something that would just take a while until it is finally there?
